requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
pytest==7.4.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
pylint==3.0.3
torch>=2.0.0
transformers>=4.30.0
datasets>=2.11.0
tqdm>=4.65.0
numpy>=1.24.0
scikit-learn>=1.2.0
pandas>=1.5.0